worker-shopify-orders/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.2.0
jinja2==3.1.4
aiofiles==24.1.0

# External API clients
openai==1.51.2
stripe==11.1.0
twilio==9.3.6

# Additional utilities
httpx==0.27.2
python-multipart==0.0.12

# Image processing
Pillow==11.0.0
